update readme re: testing
I made an attempt to update the documentation regarding testing,
since this is literally the only place I could find a working
config to get jest and preact to play nicely.

I also wasn't sure if sinon was still being used alongside jest's
fake timers, so can add that back into the readme if it is.